% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prod_qmat1.R
\name{prod_qmat1}
\alias{prod_qmat1}
\title{Produces weight matrix for special LLTM application (e.g. position effects)}
\usage{
prod_qmat1(virtdat, nr_itemstem, nr_pos)
}
\arguments{
\item{virtdat}{data matrix with "virtual items" (e.g. resulting from function virtblock), data matrix has to be 0/1 matrix and colnames must have the form 'it30_pos4' that is the number of item and the number of position}

\item{nr_itemstem}{number of item stems}

\item{nr_pos}{number of different positions}
}
\value{
qmat weight matrix (warning: check for full rank!)
}
\description{
Function prod_qmat1 produces a weight matrix for the LLTM (e.g. estimable with package eRm) e.g. for the examination of position effects. Function is very restricted and assumes a data matrix with special colnames for producing the Q matrix. Warning: function does not proof Q for full rank!
}
\examples{
library(DataprepMisc)

it1 <- c(rep(1,5),rep(NA,7))
it2 <- c(rep(NA,5),rep(0,4), rep(NA,3))
it3 <- c(rep(NA,9),rep(0,3))
it4 <- c(rep(NA,5),rep(0,4), rep(NA,3))
it5 <- c(rep(1,5), rep(NA,7))
it6 <- c(rep(NA, 9), rep(1,3))
exqd <- cbind(it1,it2, it3, it4, it5, it6)
colnames(exqd) <- c("it1_pos1", "it1_pos2", "it1_pos3", "it2_pos1", "it2_pos2", "it2_pos3")

prod_qmat1(exqd, 2,3)
}
\author{
Christine Hohensinn
}

