% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/virtblock.R
\name{virtblock}
\alias{virtblock}
\title{Splits data matrix according to feature of group variable; produces 'virtual columns'}
\usage{
virtblock(data, where_items, where_group, pos_key)
}
\arguments{
\item{data}{contains data and a group variable}

\item{where_items}{index vector which indicates which items have to be split}

\item{where_group}{index of the grouping variable (scalar)}

\item{pos_key}{integer matrix; feature x items; start with 1!}
}
\value{
list of expanded data matrix with the split columns; each list element contains the new columns representing the expandation of one column in the original data matrix; colname shows which feature the new columns have
}
\description{
Function virtblock expands splits and expands a data matrix according to features of a group variable in it. Data matrix is a persons x variables matrix including a group variable. Part of the variables (items) should be split into independent variables / columns. The splitting is due to a feature of the group variable on the items variables. For example the group variable indicates booklets. These booklets have different positions for the items. Result is a data matrix with more columns than the original data matrix because those item variable that differ in the feature (position) are split into two or more columns (dependent on the level of the feature)
}
\examples{
library(DataprepMisc)

exdat_v <- matrix(sample(c(0,1),3*5, rep=TRUE),ncol=3)
colnames(exdat_v) <- paste0("it", 1:3)
group <- sample(1:3, 5, rep=TRUE)
position <- matrix(c(1,2,1,2,2,1),ncol=3)
exdat_v2 <- cbind(exdat_v, group)
exdat_v2

virtblock(exdat_v2, 1:3, 4, position)

}
\author{
Manuel Reif, Christine Hohensinn
}

