% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prod_qmat1.R
\name{prod_qmat1}
\alias{prod_qmat1}
\title{Produces weight matrix for special LLTM application (e.g. position effects)}
\usage{
prod_qmat1(virtdat)
}
\arguments{
\item{virtdat}{data matrix with "virtual items" (e.g. resulting from function virtblock), data matrix has to be 0/1 matrix and colnames must have the form 'it30_pos4' that is the number of item and the number of position}
}
\value{
qmat weight matrix (warning: check for full rank!)
}
\description{
Function prod_qmat1 produces a weight matrix for the LLTM (e.g. estimable with package eRm) e.g. for the examination of position effects. Function is very restricted and assumes a data matrix with special colnames for producing the Q matrix. Warning: function does not proof Q for full rank!
}
\author{
Christine Hohensinn
}

