% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/virtblock.R
\name{virtblock}
\alias{virtblock}
\title{Splits data matrix according to feature of group variable; produces 'virtual columns'}
\usage{
virtblock(data, where_items, where_group, pos_key)
}
\arguments{
\item{data}{contains data and a group variable}

\item{where_items}{index vector which indicates which items have to be split}

\item{where_group}{index of the grouping variable (scalar)}

\item{pos_key}{integer matrix; feature x items; start with 1!}
}
\value{
expanded data matrix with the split columns
}
\description{
Function virtblock expands splits and expands a data matrix according to features of a group variable in it. Data matrix is a persons x variables matrix including a group variable. Part of the variables (items) should be split into independent variables / columns. The splitting is due to a feature of the group variable on the items variables. For example the group variable indicates booklets. These booklets have different positions for the items. Result is a data matrix with more columns than the original data matrix because those item variable that differ in the feature (position) are split into two or more columns (dependent on the level of the feature)
}
\author{
Manuel Reif, Christine Hohensinn
}

